웹 메인화면 및 관리자화면 수정

+++ src/main/java/CheersMate/cheersmate/domain/dto/FeedbackStatisticsDTO.java
@@ -9,7 +9,7 @@ import java.util.Map;
 @AllArgsConstructor
 public class FeedbackStatisticsDTO {
     private double averageRating; // 평균 평점
-    private Map<String, Long> weatherStats; // 날씨 조건 통계
-    private Map<String, Long> emotionStats; // 감정 통계
-    private Map<String, Long> companionStats; // 동반자 통계
+    private Map<String, Double> weatherStats; // 날씨 조건 통계
+    private Map<String, Double> emotionStats; // 감정 통계
+    private Map<String, Double> companionStats; // 동반자 통계
 }
